+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+ - strip trailing slash in `gravityServerUrl` (see [#6](https://github.com/Smartesting/gravity-data-collector/issues/6))
+
 ### Security
 
 ## [1.0.20](https://github.com/Smartesting/gravity-data-collector/compare/v1.0.20...v1.0.20)